public/server/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
google-generativeai==0.3.2
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-api-python-client==2.108.0
python-dotenv==1.0.0
requests==2.31.0